Dragon renamed to DragonBuilder + improved Debug
Renamed the Dragon component to DragonBuilder. In the future, there'll be a DragonPreview component that can take a DragonConfig input.

I also improved the DragonDebugButtons. They now are generated using the STATES array, and the DragonBuilder ignores debug clicks if not in dev mode. (There should never be debug clicks when not in dev mode.)