Dragon Builder Basic States + UI Cleanup (#6)
* Split up Dragon into sub-components + added functionality

* Printing hides Header and Footer now

* Added DragonControlButtons

* Improved use of index.ts exports

* Dragon renamed to DragonBuilder + improved Debug

Renamed the Dragon component to DragonBuilder. In the future, there'll be a DragonPreview component that can take a DragonConfig input.

I also improved the DragonDebugButtons. They now are generated using the STATES array, and the DragonBuilder ignores debug clicks if not in dev mode. (There should never be debug clicks when not in dev mode.)

* Renamed STATES to BUILDER_STATES

* Slight rounding + vertically-aligned button groups

* Improved colors for general UI

* Improved Drawer keyboard navigation

* Loading animation slightly improved

* WIP animated border-color for container

The background property isn't rendered when printing. This update replaces the rainbow background of the DragonContainer with a border that cycles through the colors.

* DragonContainer border-color transitions

* DragonContainer height now animates

* DragonBuilder state transitions crossfade now!

* Added remaining Dragon State Contents

* Removed Share button for now

* Moved Builder States into dedicated folder

* Typed dispatching + basic Builder State functions

* Added more basic Dragon Config options

* Final cleanup before PR