One bug fix + some misc unneeded refactoring
Speeds needs to not be mandatory on MonsterVals or else DragonVals isn't compatible.

I'm not sure where I was going with this refactoring, but honestly I didn't need to do it.

I added age and color back to DragonVals. I guess this is so I can go from the DragonVals in PETAL_MONSTER_VALS to DragonStats more easily. Could allow me to remove the gross number of StaticDragonStatBlock instances.

There's now a PrismaticDragon type, to differentiate from the other PetalMonsters which don't exist.