Bug Fixes + content on Prismatic Dragons page (#41)
* Made MonsterVals require speeds and languages

They can be empty strings, but at least this way it'll be harder to forget about them.

* Frogs no longer have natural armor or XP

* One bug fix + some misc unneeded refactoring

Speeds needs to not be mandatory on MonsterVals or else DragonVals isn't compatible.

I'm not sure where I was going with this refactoring, but honestly I didn't need to do it.

I added age and color back to DragonVals. I guess this is so I can go from the DragonVals in PETAL_MONSTER_VALS to DragonStats more easily. Could allow me to remove the gross number of StaticDragonStatBlock instances.

There's now a PrismaticDragon type, to differentiate from the other PetalMonsters which don't exist.

* Removed child page from navigation

* Added content to the Prismatic Dragons page

* Fixed wrapping on Prismatic Dragons page

* Added copyright notice + edits on Prismatic Dragon

* Hiding MWB content on Prismatic Dragons page